Add in GEFS updates to rocoto_viewer
There are some changes to rocoto_viewer that were made elsewhere
for GEFS that are now brought back into CROW. Primary of these is
updating the XML parser from xml to lxml to allow workflow files
that are split into pieces using entities. The other changes are
abstracting the max job-length name to a variable and changing the
header string.

Refs: #15