Remove string modulo and fix assorted bugs in rocoto_viewer
All strings still using the modulo (%) operator for substitution
are replaced with more readable f-strings instead.

Some remaining instances that were using functions from produtil
that were missed in the previous commit are updated to replace
those functions. Two functions for validating a POSIX shell string
are copied and updated from produtil for use here.

Some minor bugs and typos are resolved.

Replaced the former cPickle (which is not available for python3)
calls with pickle calls.

Fixed bugs in rocotocheck related to passing an int argument instead
of a string to syscall and not defining current_check_time.

Corrected issue with rocoto_viewer not immediately exiting when the
workflow cannot be parsed (usually because lxml could not be loaded).

Refs: #13, #14